credit_engine/secondary_research.py
"""Secondary research using web search - Indian context"""
from typing import List, Dict
import requests
from bs4 import BeautifulSoup
import re

class SecondaryResearch:
    def __init__(self, api_key: str = None):
        self.api_key = api_key
        self.mca_base_url = "https://www.mca.gov.in"
        self.ecourts_base_url = "https://services.ecourts.gov.in"
    
    def research_entity(self, entity_name: str, cin: str = None, promoters: List[str] = None) -> Dict:
        """Perform web-scale research on Indian entity"""
        results = {
            "news": self._search_news(entity_name),
            "promoter_news": self._search_promoter_news(promoters or []),
            "mca_filings": self._search_mca(cin, entity_name),
            "legal_cases": self._search_ecourts(entity_name),
            "regulatory_actions": self._search_regulatory(entity_name),
            "sector_trends": self._search_sector_trends(entity_name),
            "credit_ratings": self._search_ratings(entity_name)
        }
        return results
    
    def _search_news(self, entity: str) -> List[Dict]:
        """Search Indian business news"""
        # Search Economic Times, Business Standard, Mint, etc.
        sources = [
            f"site:economictimes.indiatimes.com {entity}",
            f"site:business-standard.com {entity}",
            f"site:livemint.com {entity}"
        ]
        return []  # Integrate with Google News API or web scraping
    
    def _search_promoter_news(self, promoters: List[str]) -> List[Dict]:
        """Search news about promoters"""
        promoter_findings = []
        for promoter in promoters:
            # Search for negative news
            negative_keywords = ["fraud", "scam", "investigation", "ED", "CBI", "SEBI"]
            # Implement search logic
            pass
        return promoter_findings
    
    def _search_mca(self, cin: str, company_name: str) -> Dict:
        """Search MCA portal for company filings"""
        # Scrape or API call to MCA21
        return {
            "director_changes": [],
            "charge_modifications": [],
            "annual_returns_status": "Filed",
            "compliance_score": 0.85
        }
    
    def _search_ecourts(self, entity: str) -> List[Dict]:
        """Search e-Courts portal for litigation"""
        # Scrape e-Courts CNR search
        return []
    
    def _search_regulatory(self, entity: str) -> List[Dict]:
        """Search RBI, SEBI, NCLT orders"""
        regulatory_sources = [
            "site:rbi.org.in",
            "site:sebi.gov.in",
            "site:nclt.gov.in"
        ]
        return []
    
    def _search_sector_trends(self, entity: str) -> Dict:
        """Search for sector-specific headwinds/tailwinds"""
        # Example: "RBI new regulations NBFC", "textile sector slowdown"
        return {
            "sector": "",
            "trends": [],
            "regulatory_changes": []
        }
    
    def _search_ratings(self, entity: str) -> List[Dict]:
        """Search CRISIL, ICRA, CARE ratings"""
        return []
    
    def synthesize_findings(self, research_data: Dict) -> str:
        """Synthesize research into narrative summary"""
        summary = []
        
        # News sentiment
        news = research_data.get("news", [])
        if news:
            negative_news = [n for n in news if self._is_negative(n.get('title', ''))]
            if negative_news:
                summary.append(f"⚠️ {len(negative_news)} negative news articles found")
        
        # Promoter concerns
        promoter_news = research_data.get("promoter_news", [])
        if promoter_news:
            summary.append(f"⚠️ Promoter-related concerns: {len(promoter_news)} findings")
        
        # MCA compliance
        mca = research_data.get("mca_filings", {})
        if mca.get("compliance_score", 1) < 0.7:
            summary.append("⚠️ MCA compliance issues detected")
        
        # Litigation
        legal = research_data.get("legal_cases", [])
        if legal:
            total_claim = sum([case.get('claim_amount', 0) for case in legal])
            summary.append(f"⚠️ {len(legal)} legal cases found (Total claim: ₹{total_claim:,.0f})")
        
        # Regulatory actions
        regulatory = research_data.get("regulatory_actions", [])
        if regulatory:
            summary.append(f"⚠️ {len(regulatory)} regulatory actions/orders found")
        
        # Sector trends
        sector = research_data.get("sector_trends", {})
        if sector.get("regulatory_changes"):
            summary.append(f"ℹ️ Sector regulatory changes: {len(sector['regulatory_changes'])} identified")
        
        # Ratings
        ratings = research_data.get("credit_ratings", [])
        if ratings:
            latest_rating = ratings[0]
            summary.append(f"Credit Rating: {latest_rating.get('rating', 'N/A')} ({latest_rating.get('agency', 'N/A')})")
        
        return "\n".join(summary) if summary else "No significant findings from secondary research"
    
    def _is_negative(self, text: str) -> bool:
        """Check if text contains negative keywords"""
        negative_keywords = [
            'fraud', 'scam', 'investigation', 'default', 'npa', 'wilful defaulter',
            'insolvency', 'nclt', 'bankruptcy', 'sebi action', 'rbi penalty'
        ]
        return any(keyword in text.lower() for keyword in negative_keywords)
