requirements.txt
databricks-sql-connector>=2.9.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0
flask>=3.0.0
